+++ src/app.module.ts
@@ -1,12 +1,15 @@
 import { MiddlewareConsumer, Module } from '@nestjs/common';
-import { ConfigModule } from '@nestjs/config';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { GracefulShutdownModule } from '@tygra/nestjs-graceful-shutdown';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { RequestIdMiddleware } from './common/middlewares';
-import { getGracefulShutdownConfig } from './config';
+import { getGracefulShutdownConfig, getTypeOrmModuleOptions } from './config';
 import { getEnvFile, validate } from './core/environment';
+import { OrdersModule } from './orders/orders.module';
+import { ProductsModule } from './products/products.module';
 import { UsersModule } from './users/users.module';
 
 @Module({
@@ -17,8 +20,15 @@ import { UsersModule } from './users/users.module';
       isGlobal: true,
       validate,
     }),
+    TypeOrmModule.forRootAsync({
+      imports: [ConfigModule],
+      inject: [ConfigService],
+      useFactory: getTypeOrmModuleOptions,
+    }),
     GracefulShutdownModule.forRoot(getGracefulShutdownConfig()),
     UsersModule,
+    OrdersModule,
+    ProductsModule,
   ],
   providers: [AppService],
 })