+++ src/db/migrations/1770709521757-add-order-status-product-active.ts
@@ -1,0 +1,35 @@
+import { MigrationInterface, QueryRunner } from 'typeorm';
+
+export class AddOrderStatusProductActive1770709521757 implements MigrationInterface {
+  name = 'AddOrderStatusProductActive1770709521757';
+
+  public async down(queryRunner: QueryRunner): Promise<void> {
+    await queryRunner.query(
+      `CREATE TYPE "public"."orders_status_enum_old" AS ENUM('CANCELLED', 'CREATED', 'PAID')`,
+    );
+    await queryRunner.query(`ALTER TABLE "orders" ALTER COLUMN "status" DROP DEFAULT`);
+    await queryRunner.query(
+      `ALTER TABLE "orders" ALTER COLUMN "status" TYPE "public"."orders_status_enum_old" USING "status"::"text"::"public"."orders_status_enum_old"`,
+    );
+    await queryRunner.query(`ALTER TABLE "orders" ALTER COLUMN "status" SET DEFAULT 'CREATED'`);
+    await queryRunner.query(`DROP TYPE "public"."orders_status_enum"`);
+    await queryRunner.query(
+      `ALTER TYPE "public"."orders_status_enum_old" RENAME TO "orders_status_enum"`,
+    );
+  }
+
+  public async up(queryRunner: QueryRunner): Promise<void> {
+    await queryRunner.query(
+      `ALTER TYPE "public"."orders_status_enum" RENAME TO "orders_status_enum_old"`,
+    );
+    await queryRunner.query(
+      `CREATE TYPE "public"."orders_status_enum" AS ENUM('CANCELLED', 'CREATED', 'PAID')`,
+    );
+    await queryRunner.query(`ALTER TABLE "orders" ALTER COLUMN "status" DROP DEFAULT`);
+    await queryRunner.query(
+      `ALTER TABLE "orders" ALTER COLUMN "status" TYPE "public"."orders_status_enum" USING "status"::"text"::"public"."orders_status_enum"`,
+    );
+    await queryRunner.query(`ALTER TABLE "orders" ALTER COLUMN "status" SET DEFAULT 'CREATED'`);
+    await queryRunner.query(`DROP TYPE "public"."orders_status_enum_old"`);
+  }
+}