+++ src/orders/order.entity.ts
@@ -1,0 +1,52 @@
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  Index,
+  JoinColumn,
+  ManyToOne,
+  OneToMany,
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+} from 'typeorm';
+
+import { User } from '../users/user.entity';
+import { OrderItem } from './order-item.entity';
+
+export enum OrderStatus {
+  CANCELLED = 'CANCELLED',
+  CREATED = 'CREATED',
+  PAID = 'PAID',
+}
+
+@Entity('orders')
+@Index('IDX_orders_user_id', ['userId'])
+@Index('IDX_orders_created_at', ['createdAt'])
+export class Order {
+  @CreateDateColumn({ name: 'created_at', type: 'timestamptz' })
+  createdAt: Date;
+
+  @PrimaryGeneratedColumn('uuid')
+  id: string;
+
+  @OneToMany(() => OrderItem, (item) => item.order)
+  items: OrderItem[];
+
+  @Column({
+    default: OrderStatus.CREATED,
+    enum: OrderStatus,
+    enumName: 'orders_status_enum',
+    type: 'enum',
+  })
+  status: OrderStatus;
+
+  @UpdateDateColumn({ name: 'updated_at', type: 'timestamptz' })
+  updatedAt: Date;
+
+  @JoinColumn({ name: 'user_id' })
+  @ManyToOne(() => User, (user) => user.orders, { onDelete: 'CASCADE' })
+  user: User;
+
+  @Column({ name: 'user_id', type: 'uuid' })
+  userId: string;
+}