+++ src/orders/orders.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+import { OrderItem } from './order-item.entity';
+import { Order } from './order.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Order, OrderItem])],
+})
+export class OrdersModule {}